Print item pending
