--- MOBA/Assets/Prefabs_blapeze/UIInventory.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: b2c28df3de38bb443a6a6d9ad9abb3ea
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
