+++ MOBA/Assets/Resources/Player.prefab
@@ -15,6 +15,7 @@ GameObject:
   - component: {fileID: 4592691154818611573}
   - component: {fileID: 6728838932469448538}
   - component: {fileID: 66387207781892278}
+  - component: {fileID: -1430754768530259173}
   m_Layer: 0
   m_Name: Player
   m_TagString: Untagged
@@ -138,7 +139,6 @@ MonoBehaviour:
   isAlive: 0
   canDie: 0
   canBeDisplaced: 0
-  enemyTeams: 
   canShow: 0
   canHide: 0
   baseViewRange: 0
@@ -173,6 +173,22 @@ MonoBehaviour:
   sceneViewId: 0
   InstantiationId: 0
   isRuntimeInstantiated: 0
+--- !u!114 &-1430754768530259173
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1145891968548686272}
+  m_Enabled: 0
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 2a214a66f5ceccf44a5c546a05b757d1, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  entityIndex: 0
+  canAddPassiveCapacity: 0
+  canRemovePassiveCapacity: 0
+  teamTMP: -1
 --- !u!1 &2078933325802894963
 GameObject:
   m_ObjectHideFlags: 0
