+++ MOBA/Assets/Scripts/Entities/Champion/Champion.cs
@@ -1,5 +1,7 @@
+using System.Collections;
 using Controllers;
 using Entities.FogOfWar;
+using Entities.Inventory;
 using UnityEngine;
 
 namespace Entities.Champion
@@ -14,11 +16,52 @@ namespace Entities.Champion
 
         private FogOfWarManager fowm;
 
+        #region TEST_Iventory
+
+        /*private int teamTMP = 1;
+
+            IEnumerator Pomme()
+        {
+            //yield return new WaitUntil(() => PhotonNetwork.CountOfPlayers == 2);
+            yield return new WaitForSeconds(4);
+            EntityCollectionManager.AddEntity(this);
+            
+            UIManager.Instance.AssignInventory((int)entityIndex, teamTMP);
+        }
+        
+
         protected override void OnStart()
         {
-            fowm = FogOfWarManager.Instance;
-            fowm.allViewables.Add(entityIndex,this);
-        }
+            UIManager.ClickOnItem += RequestAddItem;
+
+            entityIndex = (uint)photonView.ViewID;
+
+            switch (entityIndex)
+            {
+                case 1001:
+                    teamTMP = 1;
+                    break;
+                case 2001:
+                    teamTMP = 2;
+                    break;
+                case 3001:
+                    teamTMP = 1;
+                    break;
+                case 4001:
+                    teamTMP = 2;
+                    break;
+            }
+
+            StartCoroutine(Pomme());
+        }        */
+    #endregion
+
+    
+    protected override void OnStart()
+    {
+        fowm = FogOfWarManager.Instance;
+        fowm.allViewables.Add(entityIndex,this);
+    }
 
         protected override void OnUpdate()
         {
