--- MOBA/Assets/TESTBLAP/Test_blapeze.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: e2f60241148c4ee44b41ac73dc0d932a
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
