--- MOBA/Assets/TESTBLAP/Test_entity.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2a214a66f5ceccf44a5c546a05b757d1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
