+++ MOBA/ProjectSettings/EditorBuildSettings.asset
@@ -23,4 +23,7 @@ EditorBuildSettings:
   - enabled: 1
     path: Assets/_Scenes/InGameScene.unity
     guid: 6292000b4dcaf834ea926447b9513cda
+  - enabled: 1
+    path: Assets/TESTBLAP/test_blap.unity
+    guid: fec6a52c734813f4bb256759dad79796
   m_configObjects: {}
