--- MOBA_clone_0/collabignore.txt
@@ -1,1 +1,0 @@
-*
